backend/requirements.txt
celery==5.3.1
Django==4.2.4
django-extensions==3.2.3
django-import-export==3.2.0
django-waffle==4.0.0
djangoql==0.17.1
djangorestframework==3.14.0
freezegun==1.2.2
gunicorn==21.2.0
ipython==8.14.0
parameterized==0.9.0
psycopg[binary]==3.1.10
structlog==23.1.0
